fix(review): 3rd Copilot review — docstring, detach X, float32 detector, normalized semantics metric
Four small, confident correctness/consistency fixes from the third review:

- engine.fit_fusion docstring: document "reichenbach" as an accepted
  symbolic_semantics value (alias of product; tests exercise it) -- the public
  contract now lists the full set.
- engine.fit_fusion: detach X before .cpu().numpy() (mirrors the y fix) so a
  grad-requiring torch.Tensor input doesn't raise "Can't call numpy() on Tensor
  that requires grad".
- detectors/kmeans_distance: coerce inputs to float32 (not platform float64) to
  match the float32 fusion pipeline -- fewer allocations, consistent dtype.
- engine.fit_fusion: record sym_module.semantics (the module's normalized,
  alias-resolved value) in metrics instead of the raw argument, so the metric
  reflects what actually ran (no whitespace/case/alias drift).

Verified: pydocstyle 0, black/ruff clean, faithful `python -m mypy src/` clean
(603 files), 16 kmeans + co-training tests pass.

+++ src/omni_mercury_engine/detectors/kmeans_distance.py
@@ -80,7 +80,7 @@ class KMeansDistanceDetector:
         """
         from omni_mercury_engine.cognitive.neural_memory_layer import KMeansClusterer
 
-        arr = np.atleast_2d(np.nan_to_num(np.asarray(X, dtype=float)))
+        arr = np.atleast_2d(np.nan_to_num(np.asarray(X, dtype=np.float32)))
         self._mean = arr.mean(axis=0)
         self._std = arr.std(axis=0)
         self._std[self._std < 1e-8] = 1.0
@@ -94,7 +94,7 @@ class KMeansDistanceDetector:
         """Per-sample distance to every centroid, ``(n_samples, n_clusters)``."""
         if self._clusterer is None or self._mean is None or self._std is None:
             raise RuntimeError("KMeansDistanceDetector must be fit before use")
-        arr = np.atleast_2d(np.nan_to_num(np.asarray(X, dtype=float)))
+        arr = np.atleast_2d(np.nan_to_num(np.asarray(X, dtype=np.float32)))
         scaled = (arr - self._mean) / self._std
         dists = np.asarray(self._clusterer.get_cluster_distances(scaled), dtype=np.float32)
         if dists.ndim != 2 or dists.shape[0] != len(arr):
